Revert actorsystem 24.3 to 24.2 for 24-3-7-hotfix (#8269)
