+++ ydb/core/mind/local.cpp
@@ -109,7 +109,6 @@ class TLocalNodeRegistrar : public TActorBootstrapped<TLocalNodeRegistrar> {
     ui64 UserPoolUsage = 0; // (usage uS x threads) / sec
     ui64 MemUsage = 0;
     ui64 MemLimit = 0;
-    ui64 CpuLimit = 0; // PotentialMaxThreadCount of UserPool
     double NodeUsage = 0;
 
     bool SentDrainNode = false;
@@ -273,28 +272,28 @@ class TLocalNodeRegistrar : public TActorBootstrapped<TLocalNodeRegistrar> {
         HandlePipeDestroyed(ctx);
     }
 
-    void FillResourceMaximum(NKikimrTabletBase::TMetrics* record) {
-        record->CopyFrom(ResourceLimit);
-        if (!record->HasCPU()) {
-            if (CpuLimit != 0) {
-                record->SetCPU(CpuLimit);
-            }
-        }
-        if (!record->HasMemory()) {
-            if (MemLimit != 0) {
-                record->SetMemory(MemLimit);
-            } else {
-                record->SetMemory(NSystemInfo::TotalMemorySize());
-            }
-        }
-    }
-
     void SendStatusOk(const TActorContext &ctx) {
         LOG_DEBUG_S(ctx, NKikimrServices::LOCAL, "TLocalNodeRegistrar SendStatusOk");
         TAutoPtr<TEvLocal::TEvStatus> eventStatus = new TEvLocal::TEvStatus(TEvLocal::TEvStatus::StatusOk);
         auto& record = eventStatus->Record;
         record.SetStartTime(StartTime.GetValue());
-        FillResourceMaximum(record.MutableResourceMaximum());
+        record.MutableResourceMaximum()->CopyFrom(ResourceLimit);
+        if (!record.GetResourceMaximum().HasCPU()) {
+            TExecutorPoolStats poolStats;
+            TVector<TExecutorThreadStats> statsCopy;
+            TVector<TExecutorThreadStats> sharedStatsCopy;
+            ctx.ExecutorThread.ActorSystem->GetPoolStats(AppData()->UserPoolId, poolStats, statsCopy, sharedStatsCopy);
+            if (!statsCopy.empty()) {
+                record.MutableResourceMaximum()->SetCPU(poolStats.CurrentThreadCount * 1000000);
+            }
+        }
+        if (!record.GetResourceMaximum().HasMemory()) {
+            if (MemLimit != 0) {
+                record.MutableResourceMaximum()->SetMemory(MemLimit);
+            } else {
+                record.MutableResourceMaximum()->SetMemory(NSystemInfo::TotalMemorySize());
+            }
+        }
         NTabletPipe::SendData(ctx, HivePipeClient, eventStatus.Release());
     }
 
@@ -588,7 +587,6 @@ class TLocalNodeRegistrar : public TActorBootstrapped<TLocalNodeRegistrar> {
                 record.MutableTotalResourceUsage()->SetMemory(MemUsage);
             }
             record.SetTotalNodeUsage(NodeUsage);
-            FillResourceMaximum(record.MutableResourceMaximum());
             NTabletPipe::SendData(ctx, HivePipeClient, event.Release());
             SendTabletMetricsTime = ctx.Now();
         } else {
@@ -651,8 +649,7 @@ class TLocalNodeRegistrar : public TActorBootstrapped<TLocalNodeRegistrar> {
             const NKikimrWhiteboard::TSystemStateInfo& info = record.GetSystemStateInfo(0);
             if (static_cast<ui32>(info.PoolStatsSize()) > AppData()->UserPoolId) {
                 const auto& poolStats(info.GetPoolStats(AppData()->UserPoolId));
-                CpuLimit = poolStats.limit() * 1'000'000; // microseconds
-                UserPoolUsage = poolStats.usage() * CpuLimit; // microseconds
+                UserPoolUsage = poolStats.usage() * poolStats.threads() * 1000000; // uS
             }
 
             // Note: we use allocated memory because MemoryUsed(AnonRSS) has lag
