+++ ydb/library/actors/actor_type/common.h
@@ -29,17 +29,6 @@ enum class EInternalActorType {
     NAMESERVICE,
     DNS_RESOLVER,
     INTERCONNECT_PROXY_WRAPPER,
-    ACTOR_COROUTINE,
+    ACTOR_COROUTINE
 };
-
-enum class EInternalActorSystemActivity {
-    ACTOR_SYSTEM_SPIN,
-    ACTOR_SYSTEM_HARMONIZER,
-    ACTOR_SYSTEM_SEND,
-    ACTOR_SYSTEM_REGISTER,
-    ACTOR_SYSTEM_GET_ACTIVATION,
-    ACTOR_SYSTEM_GET_ACTIVATION_FROM_QUEUE,
-    ACTOR_SYSTEM_WAKE_UP,
-};
-
 }
