+++ ydb/library/actors/core/actorsystem.cpp
@@ -1,5 +1,4 @@
 #include "defs.h"
-#include "activity_guard.h"
 #include "actorsystem.h"
 #include "callstack.h"
 #include "cpu_manager.h"
@@ -97,7 +96,6 @@ namespace NActors {
         if (Y_UNLIKELY(!ev))
             return false;
 
-        TInternalActorTypeGuard<EInternalActorSystemActivity::ACTOR_SYSTEM_SEND, false> activityGuard;
 #ifdef USE_ACTOR_CALLSTACK
         ev->Callstack.TraceIfEmpty();
 #endif
@@ -338,13 +336,4 @@ namespace NActors {
         CpuManager->Cleanup();
         Scheduler.Destroy();
     }
-
-    void TActorSystem::GetExecutorPoolState(i16 poolId, TExecutorPoolState &state) const {
-        CpuManager->GetExecutorPoolState(poolId, state);
-    }
-
-    void TActorSystem::GetExecutorPoolStates(std::vector<TExecutorPoolState> &states) const {
-        CpuManager->GetExecutorPoolStates(states);
-    }
-
 }
