+++ ydb/library/actors/core/actorsystem.h
@@ -306,8 +306,5 @@ namespace NActors {
             return CpuManager->GetBasicExecutorPools();
         }
 
-        void GetExecutorPoolState(i16 poolId, TExecutorPoolState &state) const;
-        void GetExecutorPoolStates(std::vector<TExecutorPoolState> &states) const;
-
     };
 }
