+++ ydb/library/actors/core/cpu_manager.h
@@ -5,7 +5,6 @@
 #include "harmonizer.h"
 #include "executor_pool.h"
 #include "executor_pool_shared.h"
-#include "mon_stats.h"
 #include <memory>
 
 namespace NActors {
@@ -48,8 +47,6 @@ namespace NActors {
         }
 
         void GetPoolStats(ui32 poolId, TExecutorPoolStats& poolStats, TVector<TExecutorThreadStats>& statsCopy, TVector<TExecutorThreadStats>& sharedStatsCopy) const;
-        void GetExecutorPoolState(i16 poolId, TExecutorPoolState &state) const;
-        void GetExecutorPoolStates(std::vector<TExecutorPoolState> &states) const;
 
         THarmonizerStats GetHarmonizerStats() const {
             if (Harmonizer) {
