+++ ydb/library/actors/core/event_pb.cpp
@@ -1,8 +1,5 @@
 #include "event_pb.h"
 
-// enable only when patch with this macro was successfully deployed
-#define USE_EXTENDED_PAYLOAD_FORMAT 0
-
 namespace NActors {
     bool TRopeStream::Next(const void** data, int* size) {
         *data = Iter.ContiguousData();
@@ -224,213 +221,4 @@ namespace NActors {
         Buffers->Append(*s);
         return true;
     }
-
-    size_t SerializeNumber(size_t num, char *buffer) {
-        char *begin = buffer;
-        do {
-            *buffer++ = (num & 0x7F) | (num >= 128 ? 0x80 : 0x00);
-            num >>= 7;
-        } while (num);
-        return buffer - begin;
-    }
-
-    size_t DeserializeNumber(TRope::TConstIterator& iter, ui64& size) {
-        size_t res = 0;
-        size_t offset = 0;
-        for (;;) {
-            if (!iter.Valid()) {
-                return Max<size_t>();
-            }
-            const char byte = *iter.ContiguousData();
-            iter += 1;
-            --size;
-            res |= (static_cast<size_t>(byte) & 0x7F) << offset;
-            offset += 7;
-            if (!(byte & 0x80)) {
-                break;
-            }
-        }
-        return res;
-    }
-
-    bool SerializeToArcadiaStreamImpl(TChunkSerializer* chunker, const TVector<TRope> &payload) {
-        // serialize payload first
-        if (payload) {
-            void *data;
-            int size = 0;
-            auto append = [&](const char *p, size_t len) {
-                while (len) {
-                    if (size) {
-                        const size_t numBytesToCopy = std::min<size_t>(size, len);
-                        memcpy(data, p, numBytesToCopy);
-                        data = static_cast<char*>(data) + numBytesToCopy;
-                        size -= numBytesToCopy;
-                        p += numBytesToCopy;
-                        len -= numBytesToCopy;
-                    } else if (!chunker->Next(&data, &size)) {
-                        return false;
-                    }
-                }
-                return true;
-            };
-            auto appendNumber = [&](size_t number) {
-                char buf[MaxNumberBytes];
-                return append(buf, SerializeNumber(number, buf));
-            };
-
-#if USE_EXTENDED_PAYLOAD_FORMAT
-            char marker = ExtendedPayloadMarker;
-            append(&marker, 1);
-            if (!appendNumber(payload.size())) {
-                return false;
-            }
-            for (const TRope& rope : payload) {
-                if (!appendNumber(rope.GetSize())) {
-                    return false;
-                }
-            }
-            if (size) {
-                chunker->BackUp(std::exchange(size, 0));
-            }
-            for (const TRope& rope : payload) {
-                if (!chunker->WriteRope(&rope)) {
-                    return false;
-                }
-            }
-#else
-            char marker = PayloadMarker;
-            append(&marker, 1);
-            if (!appendNumber(payload.size())) {
-                return false;
-            }
-            for (const TRope& rope : payload) {
-                if (!appendNumber(rope.GetSize())) {
-                    return false;
-                }
-                if (rope) {
-                    if (size) {
-                        chunker->BackUp(std::exchange(size, 0));
-                    }
-                    if (!chunker->WriteRope(&rope)) {
-                        return false;
-                    }
-                }
-            }
-            if (size) {
-                chunker->BackUp(size);
-            }
-#endif
-        }
-
-        return true;
-    }
-
-
-    void ParseExtendedFormatPayload(TRope::TConstIterator &iter, size_t &size, TVector<TRope> &payload, size_t &totalPayloadSize)
-    {
-        // check marker
-        if (!iter.Valid() || (*iter.ContiguousData() != PayloadMarker && *iter.ContiguousData() != ExtendedPayloadMarker)) {
-            Y_ABORT("invalid event");
-        }
-
-        const bool dataIsSeparate = *iter.ContiguousData() == ExtendedPayloadMarker; // ropes go after sizes
-
-        auto fetchRope = [&](size_t len) {
-            TRope::TConstIterator begin = iter;
-            iter += len;
-            size -= len;
-            payload.emplace_back(begin, iter);
-            totalPayloadSize += len;
-        };
-
-        // skip marker
-        iter += 1;
-        --size;
-        // parse number of payload ropes
-        size_t numRopes = DeserializeNumber(iter, size);
-        if (numRopes == Max<size_t>()) {
-            Y_ABORT("invalid event");
-        }
-        TStackVec<size_t, 16> ropeLens;
-        if (dataIsSeparate) {
-            ropeLens.reserve(numRopes);
-
-        }
-        while (numRopes--) {
-            // parse length of the rope
-            const size_t len = DeserializeNumber(iter, size);
-            if (len == Max<size_t>() || size < len) {
-                Y_ABORT("invalid event len# %zu size# %" PRIu64, len, size);
-            }
-            // extract the rope
-            if (dataIsSeparate) {
-                ropeLens.push_back(len);
-            } else {
-                fetchRope(len);
-            }
-        }
-        for (size_t len : ropeLens) {
-            fetchRope(len);
-        }
-    }
-
-    ui32 CalculateSerializedSizeImpl(const TVector<TRope> &payload, size_t totalPayloadSize, ssize_t recordSize) {
-        // TODO: calc totalPayloadSize via payload? 
-        ssize_t result = recordSize;
-        if (result >= 0 && payload) {
-            ++result; // marker
-            char buf[MaxNumberBytes];
-            result += SerializeNumber(payload.size(), buf);
-            for (const TRope& rope : payload) {
-                result += SerializeNumber(rope.GetSize(), buf);
-            }
-            result += totalPayloadSize;
-        }
-        return result;
-    }
-
-    TEventSerializationInfo CreateSerializationInfoImpl(size_t preserializedSize, bool allowExternalDataChannel, const TVector<TRope> &payload, size_t totalPayloadSize, ssize_t recordSize) {
-            TEventSerializationInfo info;
-            info.IsExtendedFormat = static_cast<bool>(payload);
-
-            if (allowExternalDataChannel) {
-                if (payload) {
-                    char temp[MaxNumberBytes];
-#if USE_EXTENDED_PAYLOAD_FORMAT
-                    size_t headerLen = 1 + SerializeNumber(payload.size(), temp);
-                    for (const TRope& rope : payload) {
-                        headerLen += SerializeNumber(rope.size(), temp);
-                    }
-                    info.Sections.push_back(TEventSectionInfo{0, headerLen, 0, 0, true});
-                    for (const TRope& rope : payload) {
-                        info.Sections.push_back(TEventSectionInfo{0, rope.size(), 0, 0, false});
-                    }
-#else
-                    info.Sections.push_back(TEventSectionInfo{0, 1 + SerializeNumber(payload.size(), temp), 0, 0, true}); // payload marker and rope count
-                    for (const TRope& rope : payload) {
-                        const size_t ropeSize = rope.GetSize();
-                        info.Sections.back().Size += SerializeNumber(ropeSize, temp);
-                        info.Sections.push_back(TEventSectionInfo{0, ropeSize, 0, 0, false}); // data as a separate section
-                    }
-#endif
-                }
-
-                const size_t byteSize = Max<ssize_t>(0, recordSize) + preserializedSize;
-                info.Sections.push_back(TEventSectionInfo{0, byteSize, 0, 0, true}); // protobuf itself
-
-#ifdef NDEBUG
-                Y_UNUSED(totalPayloadSize);
-#else
-                size_t total = 0;
-                for (const auto& section : info.Sections) {
-                    total += section.Size;
-                }
-                size_t serialized = CalculateSerializedSizeImpl(payload, totalPayloadSize, recordSize);
-                Y_ABORT_UNLESS(total == serialized, "total# %zu serialized# %zu byteSize# %zd payload.size# %zu", total,
-                    serialized, byteSize, payload.size());
-#endif
-            }
-
-            return info;
-        }
 }
