+++ ydb/library/actors/core/event_pb.h
@@ -14,6 +14,9 @@
 #include <array>
 #include <span>
 
+// enable only when patch with this macro was successfully deployed
+#define USE_EXTENDED_PAYLOAD_FORMAT 0
+
 namespace NActorsProto {
     class TActorId;
 } // NActorsProto
@@ -143,15 +146,6 @@ namespace NActors {
     };
 
     static const size_t EventMaxByteSize = 140 << 20; // (140MB)
-    static constexpr char ExtendedPayloadMarker = 0x06;
-    static constexpr char PayloadMarker = 0x07;
-    static constexpr size_t MaxNumberBytes = (sizeof(size_t) * CHAR_BIT + 6) / 7;
-
-    void ParseExtendedFormatPayload(TRope::TConstIterator &iter, size_t &size, TVector<TRope> &payload, size_t &totalPayloadSize);
-    size_t SerializeNumber(size_t num, char *buffer);
-    bool SerializeToArcadiaStreamImpl(TChunkSerializer* chunker, const TVector<TRope> &payload);
-    ui32 CalculateSerializedSizeImpl(const TVector<TRope> &payload, size_t totalPayloadSize, ssize_t recordSize);
-    TEventSerializationInfo CreateSerializationInfoImpl(size_t preserializedSize, bool allowExternalDataChannel, const TVector<TRope> &payload, size_t totalPayloadSize, ssize_t recordSize);
 
     template <typename TEv, typename TRecord /*protobuf record*/, ui32 TEventType, typename TRecHolder>
     class TEventPBBase: public TEventBase<TEv, TEventType> , public TRecHolder {
@@ -194,14 +188,21 @@ namespace NActors {
         }
 
         bool SerializeToArcadiaStream(TChunkSerializer* chunker) const override {
-            if (!SerializeToArcadiaStreamImpl(chunker, Payload)) {
-                return false;
-            }
-            return Record.SerializeToZeroCopyStream(chunker);
+            return SerializeToArcadiaStreamImpl(chunker, TString());
         }
 
         ui32 CalculateSerializedSize() const override {
-            return CalculateSerializedSizeImpl(Payload, GetTotalPayloadSize(), Record.ByteSize());
+            ssize_t result = Record.ByteSize();
+            if (result >= 0 && Payload) {
+                ++result; // marker
+                char buf[MaxNumberBytes];
+                result += SerializeNumber(Payload.size(), buf);
+                for (const TRope& rope : Payload) {
+                    result += SerializeNumber(rope.GetSize(), buf);
+                }
+                result += TotalPayloadSize;
+            }
+            return result;
         }
 
         static IEventBase* Load(TEventSerializedData *input) {
@@ -213,7 +214,49 @@ namespace NActors {
                 ui64 size = input->GetSize();
 
                 if (const auto& info = input->GetSerializationInfo(); info.IsExtendedFormat) {
-                    ParseExtendedFormatPayload(iter, size, ev->Payload, ev->TotalPayloadSize);
+                    // check marker
+                    if (!iter.Valid() || (*iter.ContiguousData() != PayloadMarker && *iter.ContiguousData() != ExtendedPayloadMarker)) {
+                        Y_ABORT("invalid event");
+                    }
+
+                    const bool dataIsSeparate = *iter.ContiguousData() == ExtendedPayloadMarker; // ropes go after sizes
+
+                    auto fetchRope = [&](size_t len) {
+                        TRope::TConstIterator begin = iter;
+                        iter += len;
+                        size -= len;
+                        ev->Payload.emplace_back(begin, iter);
+                        ev->TotalPayloadSize += len;
+                    };
+
+                    // skip marker
+                    iter += 1;
+                    --size;
+                    // parse number of payload ropes
+                    size_t numRopes = DeserializeNumber(iter, size);
+                    if (numRopes == Max<size_t>()) {
+                        Y_ABORT("invalid event");
+                    }
+                    TStackVec<size_t, 16> ropeLens;
+                    if (dataIsSeparate) {
+                        ropeLens.reserve(numRopes);
+                    }
+                    while (numRopes--) {
+                        // parse length of the rope
+                        const size_t len = DeserializeNumber(iter, size);
+                        if (len == Max<size_t>() || size < len) {
+                            Y_ABORT("invalid event len# %zu size# %" PRIu64, len, size);
+                        }
+                        // extract the rope
+                        if (dataIsSeparate) {
+                            ropeLens.push_back(len);
+                        } else {
+                            fetchRope(len);
+                        }
+                    }
+                    for (size_t len : ropeLens) {
+                        fetchRope(len);
+                    }
                 }
 
                 // parse the protobuf
@@ -242,7 +285,7 @@ namespace NActors {
         }
 
         TEventSerializationInfo CreateSerializationInfo() const override {
-            return CreateSerializationInfoImpl(0, static_cast<const TEv&>(*this).AllowExternalDataChannel(), GetPayload(), GetTotalPayloadSize(), Record.ByteSize());
+            return CreateSerializationInfoImpl(0);
         }
 
         bool AllowExternalDataChannel() const {
@@ -267,10 +310,6 @@ namespace NActors {
             return Payload[id];
         }
 
-        const TVector<TRope>& GetPayload() const {
-            return Payload;
-        }
-
         ui32 GetPayloadCount() const {
             return Payload.size();
         }
@@ -278,14 +317,181 @@ namespace NActors {
         void StripPayload() {
             Payload.clear();
             TotalPayloadSize = 0;
-        } 
+        }
 
-        size_t GetTotalPayloadSize() const {
-            return TotalPayloadSize;
+    protected:
+        TEventSerializationInfo CreateSerializationInfoImpl(size_t preserializedSize) const {
+            TEventSerializationInfo info;
+            info.IsExtendedFormat = static_cast<bool>(Payload);
+
+            if (static_cast<const TEv&>(*this).AllowExternalDataChannel()) {
+                if (Payload) {
+                    char temp[MaxNumberBytes];
+#if USE_EXTENDED_PAYLOAD_FORMAT
+                    size_t headerLen = 1 + SerializeNumber(Payload.size(), temp);
+                    for (const TRope& rope : Payload) {
+                        headerLen += SerializeNumber(rope.size(), temp);
+                    }
+                    info.Sections.push_back(TEventSectionInfo{0, headerLen, 0, 0, true});
+                    for (const TRope& rope : Payload) {
+                        info.Sections.push_back(TEventSectionInfo{0, rope.size(), 0, 0, false});
+                    }
+#else
+                    info.Sections.push_back(TEventSectionInfo{0, 1 + SerializeNumber(Payload.size(), temp), 0, 0, true}); // payload marker and rope count
+                    for (const TRope& rope : Payload) {
+                        const size_t ropeSize = rope.GetSize();
+                        info.Sections.back().Size += SerializeNumber(ropeSize, temp);
+                        info.Sections.push_back(TEventSectionInfo{0, ropeSize, 0, 0, false}); // data as a separate section
+                    }
+#endif
+                }
+
+                const size_t byteSize = Max<ssize_t>(0, Record.ByteSize()) + preserializedSize;
+                info.Sections.push_back(TEventSectionInfo{0, byteSize, 0, 0, true}); // protobuf itself
+
+#ifndef NDEBUG
+                size_t total = 0;
+                for (const auto& section : info.Sections) {
+                    total += section.Size;
+                }
+                size_t serialized = CalculateSerializedSize();
+                Y_ABORT_UNLESS(total == serialized, "total# %zu serialized# %zu byteSize# %zd Payload.size# %zu", total,
+                    serialized, byteSize, Payload.size());
+#endif
+            }
+
+            return info;
+        }
+
+        bool SerializeToArcadiaStreamImpl(TChunkSerializer* chunker, const TString& preserialized) const {
+            // serialize payload first
+            if (Payload) {
+                void *data;
+                int size = 0;
+                auto append = [&](const char *p, size_t len) {
+                    while (len) {
+                        if (size) {
+                            const size_t numBytesToCopy = std::min<size_t>(size, len);
+                            memcpy(data, p, numBytesToCopy);
+                            data = static_cast<char*>(data) + numBytesToCopy;
+                            size -= numBytesToCopy;
+                            p += numBytesToCopy;
+                            len -= numBytesToCopy;
+                        } else if (!chunker->Next(&data, &size)) {
+                            return false;
+                        }
+                    }
+                    return true;
+                };
+                auto appendNumber = [&](size_t number) {
+                    char buf[MaxNumberBytes];
+                    return append(buf, SerializeNumber(number, buf));
+                };
+
+#if USE_EXTENDED_PAYLOAD_FORMAT
+                char marker = ExtendedPayloadMarker;
+                append(&marker, 1);
+                if (!appendNumber(Payload.size())) {
+                    return false;
+                }
+                for (const TRope& rope : Payload) {
+                    if (!appendNumber(rope.GetSize())) {
+                        return false;
+                    }
+                }
+                if (size) {
+                    chunker->BackUp(std::exchange(size, 0));
+                }
+                for (const TRope& rope : Payload) {
+                    if (!chunker->WriteRope(&rope)) {
+                        return false;
+                    }
+                }
+#else
+                char marker = PayloadMarker;
+                append(&marker, 1);
+                if (!appendNumber(Payload.size())) {
+                    return false;
+                }
+                for (const TRope& rope : Payload) {
+                    if (!appendNumber(rope.GetSize())) {
+                        return false;
+                    }
+                    if (rope) {
+                        if (size) {
+                            chunker->BackUp(std::exchange(size, 0));
+                        }
+                        if (!chunker->WriteRope(&rope)) {
+                            return false;
+                        }
+                    }
+                }
+                if (size) {
+                    chunker->BackUp(size);
+                }
+#endif
+            }
+
+            if (preserialized && !chunker->WriteString(&preserialized)) {
+                return false;
+            }
+
+            return Record.SerializeToZeroCopyStream(chunker);
         }
 
     protected:
         mutable size_t CachedByteSize = 0;
+
+        static constexpr char ExtendedPayloadMarker = 0x06;
+        static constexpr char PayloadMarker = 0x07;
+        static constexpr size_t MaxNumberBytes = (sizeof(size_t) * CHAR_BIT + 6) / 7;
+
+        static size_t SerializeNumber(size_t num, char *buffer) {
+            char *begin = buffer;
+            do {
+                *buffer++ = (num & 0x7F) | (num >= 128 ? 0x80 : 0x00);
+                num >>= 7;
+            } while (num);
+            return buffer - begin;
+        }
+
+        static size_t DeserializeNumber(const char **ptr, const char *end) {
+            const char *p = *ptr;
+            size_t res = 0;
+            size_t offset = 0;
+            for (;;) {
+                if (p == end) {
+                    return Max<size_t>();
+                }
+                const char byte = *p++;
+                res |= (static_cast<size_t>(byte) & 0x7F) << offset;
+                offset += 7;
+                if (!(byte & 0x80)) {
+                    break;
+                }
+            }
+            *ptr = p;
+            return res;
+        }
+
+        static size_t DeserializeNumber(TRope::TConstIterator& iter, ui64& size) {
+            size_t res = 0;
+            size_t offset = 0;
+            for (;;) {
+                if (!iter.Valid()) {
+                    return Max<size_t>();
+                }
+                const char byte = *iter.ContiguousData();
+                iter += 1;
+                --size;
+                res |= (static_cast<size_t>(byte) & 0x7F) << offset;
+                offset += 7;
+                if (!(byte & 0x80)) {
+                    break;
+                }
+            }
+            return res;
+        }
     };
 
     // Protobuf record not using arena
@@ -419,16 +625,7 @@ namespace NActors {
         }
 
         bool SerializeToArcadiaStream(TChunkSerializer* chunker) const override {
-            if (!SerializeToArcadiaStreamImpl(chunker, TBase::GetPayload())) {
-                return false;
-            }
-            
-            if (PreSerializedData && !chunker->WriteString(&PreSerializedData)) {
-                return false;
-            }
-
-            return Record.SerializeToZeroCopyStream(chunker);
-            
+            return TBase::SerializeToArcadiaStreamImpl(chunker, PreSerializedData);
         }
 
         ui32 CalculateSerializedSize() const override {
@@ -444,7 +641,7 @@ namespace NActors {
         }
 
         TEventSerializationInfo CreateSerializationInfo() const override {
-            return CreateSerializationInfoImpl(PreSerializedData.size(), static_cast<const TEv&>(*this).AllowExternalDataChannel(), TBase::GetPayload(), TBase::GetTotalPayloadSize(), Record.ByteSize());
+            return TBase::CreateSerializationInfoImpl(PreSerializedData.size());
         }
     };
