+++ ydb/library/actors/core/executor_pool.h
@@ -9,7 +9,6 @@ namespace NActors {
     struct TMailboxHeader;
     struct TWorkerContext;
     struct TExecutorPoolStats;
-    struct TExecutorPoolState;
     struct TExecutorThreadStats;
     class TExecutorPoolJail;
     class ISchedulerCookie;
@@ -107,10 +106,6 @@ namespace NActors {
             // TODO: make pure virtual and override everywhere
             Y_UNUSED(poolStats);
             Y_UNUSED(statsCopy);
-        }
-
-        virtual void GetExecutorPoolState(TExecutorPoolState &poolState) const {
-            Y_UNUSED(poolState);
         }
 
         virtual TString GetName() const {
