+++ ydb/library/actors/core/executor_pool_basic.h
@@ -151,13 +151,14 @@ namespace NActors {
         const TString PoolName;
         const TDuration TimePerMailbox;
         const ui32 EventsPerMailbox;
+        EASProfile ActorSystemProfile;
 
         const int RealtimePriority;
 
-        TAtomic ThreadUtilization = 0;
-        TAtomic MaxUtilizationCounter = 0;
-        TAtomic MaxUtilizationAccumulator = 0;
-        TAtomic WrongWakenedThreadCount = 0;
+        TAtomic ThreadUtilization;
+        TAtomic MaxUtilizationCounter;
+        TAtomic MaxUtilizationAccumulator;
+        TAtomic WrongWakenedThreadCount;
         std::atomic<ui64> SpinningTimeUs;
 
         TAtomic ThreadCount;
@@ -207,7 +208,6 @@ namespace NActors {
             }
         };
 
-        const EASProfile ActorSystemProfile;
         static constexpr TDuration DEFAULT_TIME_PER_MAILBOX = TBasicExecutorPoolConfig::DEFAULT_TIME_PER_MAILBOX;
         static constexpr ui32 DEFAULT_EVENTS_PER_MAILBOX = TBasicExecutorPoolConfig::DEFAULT_EVENTS_PER_MAILBOX;
 
@@ -251,7 +251,6 @@ namespace NActors {
         void Shutdown() override;
 
         void GetCurrentStats(TExecutorPoolStats& poolStats, TVector<TExecutorThreadStats>& statsCopy) const override;
-        void GetExecutorPoolState(TExecutorPoolState &poolState) const override;
         TString GetName() const override {
             return PoolName;
         }
