+++ ydb/library/actors/core/executor_thread_ctx.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "defs.h"
-#include "activity_guard.h"
 #include "executor_thread.h"
 #include "thread_context.h"
 
@@ -22,7 +21,7 @@ namespace NActors {
     };
 
     struct TGenericExecutorThreadCtx {
-        std::unique_ptr<TGenericExecutorThread> Thread;
+        TAutoPtr<TGenericExecutorThread> Thread;
 
     protected:
         friend class TIOExecutorPool;
@@ -32,8 +31,6 @@ namespace NActors {
         std::atomic<ui64> WaitingFlag = static_cast<ui64>(EThreadState::None);
 
     public:
-        ~TGenericExecutorThreadCtx(); // in executor_thread.cpp
-
         ui64 StartWakingTs = 0;
 
         ui64 GetStateInt() {
@@ -61,13 +58,11 @@ namespace NActors {
 
         template <typename TDerived, typename TWaitState>
         void Spin(ui64 spinThresholdCycles, std::atomic<bool> *stopFlag) {
+            ui64 start = GetCycleCountFast();
             bool doSpin = true;
-            NHPTimer::STime start = GetCycleCountFast();
-            TInternalActorTypeGuard<EInternalActorSystemActivity::ACTOR_SYSTEM_SPIN> activityGuard(start);
             while (true) {
                 for (ui32 j = 0; doSpin && j < 12; ++j) {
-                    NHPTimer::STime hpnow = GetCycleCountFast();
-                    if (hpnow >= i64(start + spinThresholdCycles)) {
+                    if (GetCycleCountFast() >= (start + spinThresholdCycles)) {
                         doSpin = false;
                         break;
                     }
@@ -103,8 +98,8 @@ namespace NActors {
 
             NHPTimer::STime hpnow = GetCycleCountFast();
             NHPTimer::STime hpprev = TlsThreadContext->UpdateStartOfProcessingEventTS(hpnow);
-            ui32 prevActivity = TlsThreadContext->ElapsingActorActivity.exchange(Max<ui64>(), std::memory_order_acq_rel);
-            TlsThreadContext->WorkerCtx->AddElapsedCycles(prevActivity, hpnow - hpprev);
+            TlsThreadContext->ElapsingActorActivity.store(Max<ui64>(), std::memory_order_release);
+            TlsThreadContext->WorkerCtx->AddElapsedCycles(TlsThreadContext->ActorSystemIndex, hpnow - hpprev);
             do {
                 if (WaitingPad.Park()) // interrupted
                     return true;
