+++ ydb/library/actors/core/harmonizer.cpp
@@ -4,13 +4,11 @@
 #include "executor_thread.h"
 #include "probes.h"
 
-#include "activity_guard.h"
 #include "actorsystem.h"
 #include "executor_pool_basic.h"
 #include "executor_pool_basic_feature_flags.h"
 #include "executor_pool_shared.h"
 
-#include <atomic>
 #include <ydb/library/actors/util/cpu_load_log.h>
 #include <ydb/library/actors/util/datetime.h>
 #include <ydb/library/actors/util/intrinsics.h>
@@ -203,11 +201,10 @@ struct TPoolInfo {
     TValueHistory<16> Consumed;
     TValueHistory<16> Booked;
 
-    std::atomic<float> MaxConsumedCpu = 0;
-    std::atomic<float> MinConsumedCpu = 0;
-    std::atomic<float> AvgConsumedCpu = 0;
-    std::atomic<float> MaxBookedCpu = 0;
-    std::atomic<float> MinBookedCpu = 0;
+    TAtomic MaxConsumedCpu = 0;
+    TAtomic MinConsumedCpu = 0;
+    TAtomic MaxBookedCpu = 0;
+    TAtomic MinBookedCpu = 0;
 
     std::unique_ptr<TWaitingStats<ui64>> WaitingStats;
     std::unique_ptr<TWaitingStats<double>> MovingWaitingStats;
@@ -297,7 +294,7 @@ TCpuConsumption TPoolInfo::PullStats(ui64 ts) {
     }
     TVector<TExecutorThreadStats> sharedStats;
     if (Shared) {
-        Shared->GetSharedStatsForHarmonizer(Pool->PoolId, sharedStats);
+        Shared->GetSharedStats(Pool->PoolId, sharedStats);
     }
 
     for (ui32 sharedIdx = 0; sharedIdx < SharedInfo.size(); ++sharedIdx) {
@@ -315,12 +312,11 @@ TCpuConsumption TPoolInfo::PullStats(ui64 ts) {
     }
 
     Consumed.Register(ts, acc.ConsumedUs);
-    MaxConsumedCpu.store(Consumed.GetMax() / 1'000'000, std::memory_order_relaxed);
-    MinConsumedCpu.store(Consumed.GetMin() / 1'000'000, std::memory_order_relaxed);
-    AvgConsumedCpu.store(Consumed.GetAvgPart() / 1'000'000, std::memory_order_relaxed);
+    RelaxedStore(&MaxConsumedCpu, Consumed.GetMaxInt());
+    RelaxedStore(&MinConsumedCpu, Consumed.GetMinInt());
     Booked.Register(ts, acc.BookedUs);
-    MaxBookedCpu.store(Booked.GetMax() / 1'000'000, std::memory_order_relaxed);
-    MinBookedCpu.store(Booked.GetMin() / 1'000'000, std::memory_order_relaxed);
+    RelaxedStore(&MaxBookedCpu, Booked.GetMaxInt());
+    RelaxedStore(&MinBookedCpu, Booked.GetMinInt());
     NewNotEnoughCpuExecutions = acc.NotEnoughCpuExecutions - NotEnoughCpuExecutions;
     NotEnoughCpuExecutions = acc.NotEnoughCpuExecutions;
     if (WaitingStats && BasicPool) {
@@ -362,7 +358,7 @@ private:
     std::atomic<bool> IsDisabled = false;
     TSpinLock Lock;
     std::atomic<ui64> NextHarmonizeTs = 0;
-    std::vector<std::unique_ptr<TPoolInfo>> Pools;
+    std::vector<TPoolInfo> Pools;
     std::vector<ui16> PriorityOrder;
 
     TValueHistory<16> Consumed;
@@ -407,8 +403,8 @@ double THarmonizer::Rescale(double value) const {
 
 void THarmonizer::PullStats(ui64 ts) {
     TCpuConsumption acc;
-    for (auto &pool : Pools) {
-        TCpuConsumption consumption = pool->PullStats(ts);
+    for (TPoolInfo &pool : Pools) {
+        TCpuConsumption consumption = pool.PullStats(ts);
         acc.Add(consumption);
     }
     Consumed.Register(ts, acc.ConsumedUs);
@@ -445,7 +441,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
     ui64 TotalAwakeningTime = 0;
     ui64 TotalAwakenings = 0;
     for (size_t poolIdx = 0; poolIdx < Pools.size(); ++poolIdx) {
-        TPoolInfo& pool = *Pools[poolIdx];
+        TPoolInfo& pool = Pools[poolIdx];
         if (pool.WaitingStats) {
             TotalWakingUpTime += pool.WaitingStats->WakingUpTotalTime;
             TotalWakingUps += pool.WaitingStats->WakingUpCount;
@@ -475,22 +471,20 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
     LWPROBE(WakingUpConsumption, Ts2Us(avgWakingUpTime), Ts2Us(avgWakingUpTime), Ts2Us(avgAwakeningTime), Ts2Us(realAvgAwakeningTime), Ts2Us(avgWakingUpConsumption));
 
     for (size_t poolIdx = 0; poolIdx < Pools.size(); ++poolIdx) {
-        TPoolInfo& pool = *Pools[poolIdx];
+        TPoolInfo& pool = Pools[poolIdx];
         if (!pool.BasicPool) {
             continue;
         }
-        if (pool.BasicPool->ActorSystemProfile != EASProfile::Default) {
-            if constexpr (NFeatures::TSpinFeatureFlags::CalcPerThread) {
-                pool.BasicPool->CalcSpinPerThread(avgWakingUpConsumption);
-            } else if constexpr (NFeatures::TSpinFeatureFlags::UsePseudoMovingWindow) {
-                ui64 newSpinThreshold = pool.MovingWaitingStats->CalculateGoodSpinThresholdCycles(avgWakingUpConsumption);
-                pool.BasicPool->SetSpinThresholdCycles(newSpinThreshold);
-            } else {
-                ui64 newSpinThreshold = pool.WaitingStats->CalculateGoodSpinThresholdCycles(avgWakingUpConsumption);
-                pool.BasicPool->SetSpinThresholdCycles(newSpinThreshold);
-            }
-            pool.BasicPool->ClearWaitingStats();
+        if constexpr (NFeatures::TSpinFeatureFlags::CalcPerThread) {
+            pool.BasicPool->CalcSpinPerThread(avgWakingUpConsumption);
+        } else if constexpr (NFeatures::TSpinFeatureFlags::UsePseudoMovingWindow) {
+            ui64 newSpinThreshold = pool.MovingWaitingStats->CalculateGoodSpinThresholdCycles(avgWakingUpConsumption);
+            pool.BasicPool->SetSpinThresholdCycles(newSpinThreshold);
+        } else {
+            ui64 newSpinThreshold = pool.WaitingStats->CalculateGoodSpinThresholdCycles(avgWakingUpConsumption);
+            pool.BasicPool->SetSpinThresholdCycles(newSpinThreshold);
         }
+        pool.BasicPool->ClearWaitingStats();
     }
 
     std::vector<bool> hasSharedThread(Pools.size());
@@ -515,7 +509,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
     }
 
     for (size_t poolIdx = 0; poolIdx < Pools.size(); ++poolIdx) {
-        TPoolInfo& pool = *Pools[poolIdx];
+        TPoolInfo& pool = Pools[poolIdx];
         total += pool.DefaultThreadCount;
 
         i16 currentFullThreadCount = pool.GetFullThreadCount();
@@ -597,19 +591,19 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
 
     if (needyPools.size()) {
         Sort(needyPools.begin(), needyPools.end(), [&] (i16 lhs, i16 rhs) {
-            if (Pools[lhs]->Priority != Pools[rhs]->Priority)  {
-                return Pools[lhs]->Priority > Pools[rhs]->Priority;
+            if (Pools[lhs].Priority != Pools[rhs].Priority)  {
+                return Pools[lhs].Priority > Pools[rhs].Priority;
             }
-            return Pools[lhs]->Pool->PoolId < Pools[rhs]->Pool->PoolId;
+            return Pools[lhs].Pool->PoolId < Pools[rhs].Pool->PoolId;
         });
     }
 
     if (freeHalfThread.size()) {
         Sort(freeHalfThread.begin(), freeHalfThread.end(), [&] (i16 lhs, i16 rhs) {
-            if (Pools[lhs]->Priority != Pools[rhs]->Priority)  {
-                return Pools[lhs]->Priority > Pools[rhs]->Priority;
+            if (Pools[lhs].Priority != Pools[rhs].Priority)  {
+                return Pools[lhs].Priority > Pools[rhs].Priority;
             }
-            return Pools[lhs]->Pool->PoolId < Pools[rhs]->Pool->PoolId;
+            return Pools[lhs].Pool->PoolId < Pools[rhs].Pool->PoolId;
         });
     }
 
@@ -621,7 +615,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
             // do nothing
         } else {
             for (ui16 poolIdx : PriorityOrder) {
-                TPoolInfo &pool = *Pools[poolIdx];
+                TPoolInfo &pool = Pools[poolIdx];
                 i64 threadCount = pool.GetFullThreadCount();
                 if (hasSharedThread[poolIdx] && !hasSharedThreadWhichWasNotBorrowed[poolIdx]) {
                     Shared->ReturnOwnHalfThread(poolIdx);
@@ -644,7 +638,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
         }
     } else {
         for (size_t needyPoolIdx : needyPools) {
-            TPoolInfo &pool = *Pools[needyPoolIdx];
+            TPoolInfo &pool = Pools[needyPoolIdx];
             i64 threadCount = pool.GetFullThreadCount();
             if (budget >= 1.0) {
                 if (threadCount + 1 <= pool.MaxFullThreadCount) {
@@ -676,7 +670,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
     if (budget < 1.0) {
         size_t takingAwayThreads = 0;
         for (size_t needyPoolIdx : needyPools) {
-            TPoolInfo &pool = *Pools[needyPoolIdx];
+            TPoolInfo &pool = Pools[needyPoolIdx];
             i64 threadCount = pool.GetFullThreadCount();
             sumOfAdditionalThreads -= threadCount - pool.DefaultFullThreadCount;
             if (sumOfAdditionalThreads < takingAwayThreads + 1) {
@@ -698,7 +692,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
                 break;
             }
 
-            TPoolInfo &pool = *Pools[poolIdx];
+            TPoolInfo &pool = Pools[poolIdx];
             size_t threadCount = pool.GetFullThreadCount();
             size_t additionalThreadsCount = Max<size_t>(0L, threadCount - pool.DefaultFullThreadCount);
             size_t currentTakingAwayThreads = Min(additionalThreadsCount, takingAwayThreads);
@@ -715,7 +709,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
     }
 
     for (size_t hoggishPoolIdx : hoggishPools) {
-        TPoolInfo &pool = *Pools[hoggishPoolIdx];
+        TPoolInfo &pool = Pools[hoggishPoolIdx];
         i64 threadCount = pool.GetFullThreadCount();
         if (hasBorrowedSharedThread[hoggishPoolIdx]) {
             Shared->ReturnBorrowedHalfThread(hoggishPoolIdx);
@@ -733,7 +727,7 @@ void THarmonizer::HarmonizeImpl(ui64 ts) {
     }
 
     for (size_t poolIdx = 0; poolIdx < Pools.size(); ++poolIdx) {
-        TPoolInfo& pool = *Pools[poolIdx];
+        TPoolInfo& pool = Pools[poolIdx];
         AtomicSet(pool.PotentialMaxThreadCount, std::min<i64>(pool.MaxThreadCount, pool.GetThreadCount() + budgetInt));
     }
 }
@@ -742,10 +736,10 @@ void THarmonizer::CalculatePriorityOrder() {
     PriorityOrder.resize(Pools.size());
     Iota(PriorityOrder.begin(), PriorityOrder.end(), 0);
     Sort(PriorityOrder.begin(), PriorityOrder.end(), [&] (i16 lhs, i16 rhs) {
-        if (Pools[lhs]->Priority != Pools[rhs]->Priority)  {
-            return Pools[lhs]->Priority < Pools[rhs]->Priority;
+        if (Pools[lhs].Priority != Pools[rhs].Priority)  {
+            return Pools[lhs].Priority < Pools[rhs].Priority;
         }
-        return Pools[lhs]->Pool->PoolId > Pools[rhs]->Pool->PoolId;
+        return Pools[lhs].Pool->PoolId > Pools[rhs].Pool->PoolId;
     });
 }
 
@@ -764,16 +758,12 @@ void THarmonizer::Harmonize(ui64 ts) {
     ui64 previousNextHarmonizeTs = NextHarmonizeTs.exchange(ts + Us2Ts(1'000'000ull));
     LWPROBE(TryToHarmonizeSuccess, ts, NextHarmonizeTs, previousNextHarmonizeTs);
 
-    {
-        TInternalActorTypeGuard<EInternalActorSystemActivity::ACTOR_SYSTEM_HARMONIZER> activityGuard;
-
-        if (PriorityOrder.empty()) {
-            CalculatePriorityOrder();
-        }
-
-        PullStats(ts);
-        HarmonizeImpl(ts);
+    if (PriorityOrder.empty()) {
+        CalculatePriorityOrder();
     }
+
+    PullStats(ts);
+    HarmonizeImpl(ts);
 
     Lock.Release();
 }
@@ -784,8 +774,7 @@ void THarmonizer::DeclareEmergency(ui64 ts) {
 
 void THarmonizer::AddPool(IExecutorPool* pool, TSelfPingInfo *pingInfo) {
     TGuard<TSpinLock> guard(Lock);
-    Pools.emplace_back(new TPoolInfo);
-    TPoolInfo &poolInfo = *Pools.back();
+    TPoolInfo poolInfo;
     poolInfo.Pool = pool;
     poolInfo.Shared = Shared;
     poolInfo.BasicPool = dynamic_cast<TBasicExecutorPool*>(pool);
@@ -809,6 +798,7 @@ void THarmonizer::AddPool(IExecutorPool* pool, TSelfPingInfo *pingInfo) {
         poolInfo.WaitingStats.reset(new TWaitingStats<ui64>());
         poolInfo.MovingWaitingStats.reset(new TWaitingStats<double>());
     }
+    Pools.push_back(std::move(poolInfo));
     PriorityOrder.clear();
 }
 
@@ -822,7 +812,7 @@ IHarmonizer* MakeHarmonizer(ui64 ts) {
 }
 
 TPoolHarmonizerStats THarmonizer::GetPoolStats(i16 poolId) const {
-    const TPoolInfo &pool = *Pools[poolId];
+    const TPoolInfo &pool = Pools[poolId];
     ui64 flags = RelaxedLoad(&pool.LastFlags);
     return TPoolHarmonizerStats{
         .IncreasingThreadsByNeedyState = static_cast<ui64>(RelaxedLoad(&pool.IncreasingThreadsByNeedyState)),
@@ -830,11 +820,10 @@ TPoolHarmonizerStats THarmonizer::GetPoolStats(i16 poolId) const {
         .DecreasingThreadsByStarvedState = static_cast<ui64>(RelaxedLoad(&pool.DecreasingThreadsByStarvedState)),
         .DecreasingThreadsByHoggishState = static_cast<ui64>(RelaxedLoad(&pool.DecreasingThreadsByHoggishState)),
         .DecreasingThreadsByExchange = static_cast<ui64>(RelaxedLoad(&pool.DecreasingThreadsByExchange)),
-        .MaxConsumedCpu = pool.MaxConsumedCpu.load(std::memory_order_relaxed),
-        .MinConsumedCpu = pool.MinConsumedCpu.load(std::memory_order_relaxed),
-        .AvgConsumedCpu = pool.AvgConsumedCpu.load(std::memory_order_relaxed),
-        .MaxBookedCpu = pool.MaxBookedCpu.load(std::memory_order_relaxed),
-        .MinBookedCpu = pool.MinBookedCpu.load(std::memory_order_relaxed),
+        .MaxConsumedCpu = static_cast<i64>(RelaxedLoad(&pool.MaxConsumedCpu)),
+        .MinConsumedCpu = static_cast<i64>(RelaxedLoad(&pool.MinConsumedCpu)),
+        .MaxBookedCpu = static_cast<i64>(RelaxedLoad(&pool.MaxBookedCpu)),
+        .MinBookedCpu = static_cast<i64>(RelaxedLoad(&pool.MinBookedCpu)),
         .PotentialMaxThreadCount = static_cast<i16>(RelaxedLoad(&pool.PotentialMaxThreadCount)),
         .IsNeedy = static_cast<bool>(flags & 1),
         .IsStarved = static_cast<bool>(flags & 2),
