+++ ydb/library/actors/helpers/pool_stats_collector.h
@@ -214,7 +214,6 @@ private:
         THPTimer UsageTimer;
         TString Name;
         double Threads;
-        double LimitThreads;
 
         void Init(NMonitoring::TDynamicCounters* group, const TString& poolName, ui32 threads) {
             LastElapsedSeconds = 0;
@@ -222,7 +221,6 @@ private:
             UsageTimer.Reset();
             Name = poolName;
             Threads = threads;
-            LimitThreads = threads;
 
             PoolGroup = group->GetSubgroup("execpool", poolName);
 
@@ -376,7 +374,6 @@ private:
             Y_UNUSED(stats);
 #endif
             Threads = poolStats.CurrentThreadCount;
-            LimitThreads = poolStats.PotentialMaxThreadCount;
         }
     };
